WIP - Initial start of fixing performance issues for 35011
This is a test for people to try out for a possible partial fix for 35011 and other issues. There is many layers to the the performance issues with the viewer, it is a combination of larger amount of textures allowed to be loaded at any given time, larger cache sizes for files store on local storage and number of threads and locks being used throughout the code.

This code contains a first couple of steps. There is an drop down combo box in the Preferences->Graphics->Hardware Settings which allow the user to pick form several options.

The first 3 are done currently. The first is normal behavior. The second is to use a different getFetchState method which does 2 task in one call, instead of in 2. The situation of rapid resize requests which would normally call out to the previous worker thread and get the old decoded texture. The problem is it did two sets of locks in a row on a worker which may be working on another texture and could be doing http requests. This would cause spin lock on the main thread and hitching.

The third option bypasses this request for the previous worker thread all together as it now stores a copy of the previous Raw and Aux Raw in an array on the texture itself. Allowing the texture to just re-use the previous data and to not have to go back for it again.

This is also done when a texture wants to resize before going through the process of creating a worker thread to pull the texture, it checks the array to see if a texture exists at the quality it wishes for and if so, just uses it.

There is a clean up function in with the other LLViewerTextureList update decode priority to delete all the raw textures on a given texture if they have not been updated or accessed after 30 seconds. This could be increased or added as user accessible value to give them room to increase the value if they are having yo-yo affects of texture and CPU usage.

It could be have more performance if we would bypass more of the texture system and directly create and load the texture with the commands and leave the loop that they have setup. Also, if we could use std::future as a way to kick off multiple tasks or even just to kick off and use the future.Valid() to check if it is ready and if it isn't go and process the next texture and at the end, join the results... This could be the solution to use LLMutexTryLock instead of LLMutexLock's as it is there, but the system does not handle it well when a worker comes back as not done yet. I did a bit of a test and caused some rendering issues as well as textures getting stuck in the queue. But I will take further look at that being a possible solution to the spinlocks we currently are suffering from.