Updated with some more updates to the raw list and some changes to the cache system
Added additional low quality texture level 6 and added code to add texture from fast cache to the array of raws.
Modified the trytoclearraws to leave the lower 2 levels on the list similar to how there was an embedded raw texture on the fetch texture back in 7.1.10.
Updated the cache objects to use a std::unordered_map and std::unordered_set instead of the std::map and std::set to try to help when there is a lare number of objects in the cache.
Updated the  LLViewerTextureList::updateImagesFetchTextures(F32 max_time) to not re-create the entries vector every frame and isntead store it in the class, and clear it at the end of the frame. Keeping the allocation away.
Added fix for texture cache which was not reserving the vector used for the entries being loaded from disk.
Started to work on adding a lock free queue or at least atomic values to check on the status of the worker threads without having to lock the queue of worker threads to find their status.