Updated with latest changes for performance issues
Added new Atomic state of the texture worker share between the LLViewerFetchedTexture and LLTextureFetchWorker allowing to skip over needing to lock the worker to get it's know state.
Added particle pool to speed up creating and deleting particles (We don't, just a fixed size pool, so no more need to check the number of particles as there is always a fixed size)
Added code to max the time updating imposters to a time slice instead of updating all that needed to be updated on a given frame.
Fixed issue of cache not being written to if texture was downscaled and upscaled. (This was causing  system to redownload the textures multiple times.)
Updated options in the preference to reflect all the changes. All be the 2 for Cubemaps are done implemented.
Removed a few redundant OpenGL api calls.
Started to work on new Atomic Lock-free node.

There are further improvements come, but this is the current update.