Fixed up array for the raws, and put in trap for invalid requests being sent to lltexturefetch
Changed max discard level from +2 to + 1 for the array of raws as we were storing invalid raws.
Changed the tr to clear to remove all raws and not retain any. (Could be made to be user configurable)
Added trap in the lltexturefetch dowork for invalid mDesiredDiscard values ( mDesiredDiscard < 0 || > MAX_DISCARD_VALUE) with a warning so you can put a break point and walk through the code.